backend/app/requirements.txt
alembic==1.11.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Dance==7.0.0
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
load-dotenv==0.1.0
Mako==1.2.4
MarkupSafe==2.1.3
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.1
pluggy==1.2.0
PyJWT==2.7.0
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
SQLAlchemy==2.0.18
typing_extensions==4.7.1
urllib3==2.0.3
URLObject==2.4.3
Werkzeug==2.3.6
WTForms==3.0.1
